feat(gui): profiles tab leads with user content; Desktop naming
Once a profile is saved, the tab reorders: saved profiles on
top with the native-Space bindings directly beneath them, and
the built-in Presets — a bootstrap tool — close the tab. While
no profile exists yet, Presets stay expanded on top so the
first-run view is never barren. No disclosure folding on any
section: collapsed content was tried and rejected as clutter.

Native-Space rows now read "Desktop n" (the name Mission
Control shows) — "Space n" is how KiwiDesk's own virtual
spaces read, and reusing it blurred the two systems. Binding
is dropdown-only: the draggable profile chips duplicated the
per-row dropdown, so DraggableProfile and ProfileChip are
gone. Presets split into PresetsSection.swift for file size.
