+++ Sources/KiwiDesk/Settings/Sections/ProfilesSection.swift
@@ -1,101 +1,46 @@
 import KiwiDeskCore
 import SwiftUI
 
-/// Whole App ▸ Profiles (#36/#53/#68 §3.2): the built-in
-/// per-count layouts (applyable Presets), the saved-profiles
-/// list grouped by screen count, and the native-Space bindings
-/// — "which profile applies where/when" in one place.
+/// Whole App ▸ Profiles (#36/#53/#68 §3.2): the saved-profiles
+/// list grouped by screen count, the native-Space bindings,
+/// and the built-in Presets — "which profile applies
+/// where/when" in one place. The Desktop bindings always ride
+/// directly beneath the saved profiles they reference. Presets
+/// lead only while no profile is saved yet (bootstrap, so
+/// first launch is never barren); once one exists, the user's
+/// own content takes the top and the full preset list closes
+/// the tab.
 struct ProfilesSection: View {
     @ObservedObject var model: SettingsModel
 
     var body: some View {
         ScrollView {
             VStack(alignment: .leading, spacing: 20) {
-                presetSection
-                profileSection
-                // "Which profile applies where/when" lives
-                // together (#68 §3.2): native-Space bindings
-                // are global — only offered in live editing.
-                if !model.editingStoredProfile {
-                    NativeSpacesSection(model: model)
+                if model.profileSummaries.isEmpty {
+                    PresetsSection(model: model)
+                    profileSection
+                    nativeSpaces
+                } else {
+                    profileSection
+                    nativeSpaces
+                    PresetsSection(model: model)
                 }
             }
+            .animation(
+                .default,
+                value: model.profileSummaries.isEmpty
+            )
             .padding(16)
         }
     }
 
-    // MARK: - Built-in presets (#53)
-
-    private var presetSection: some View {
-        SettingsSection("Presets") {
-            Text(
-                "Built-in layouts per screen count. Apply "
-                    + "loads one and saves it as a real, "
-                    + "editable profile — only available when "
-                    + "the connected screen count matches."
-            )
-            .font(.caption)
-            .foregroundStyle(.secondary)
-            ForEach(presetCounts, id: \.self) { count in
-                countHeader(count)
-                ForEach(
-                    StandardProfiles.layouts(for: count),
-                    id: \.name
-                ) { layout in
-                    presetRow(layout)
-                }
-            }
+    /// "Which profile applies where/when" lives together
+    /// (#68 §3.2): native-Space bindings are global — only
+    /// offered in live editing.
+    @ViewBuilder private var nativeSpaces: some View {
+        if !model.editingStoredProfile {
+            NativeSpacesSection(model: model)
         }
-    }
-
-    private var presetCounts: [Int] {
-        Array(
-            Set(StandardProfiles.all.map(\.screenCount))
-        ).sorted()
-    }
-
-    private func countHeader(_ count: Int) -> some View {
-        Text(
-            count == 1
-                ? "1 screen" : "\(count) screens"
-        )
-        .font(.caption)
-        .fontWeight(.semibold)
-        .foregroundStyle(.secondary)
-        .padding(.top, 4)
-    }
-
-    private func presetRow(
-        _ layout: StandardLayout
-    ) -> some View {
-        HStack {
-            VStack(alignment: .leading) {
-                HStack(spacing: 6) {
-                    Text(layout.name).font(.headline)
-                    if layout.isStandard {
-                        BadgeChip(label: "standard")
-                    }
-                }
-                Text(layout.summary)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                PresetThumbnail(layout: layout)
-            }
-            Spacer()
-            Button("Apply") {
-                model.applyStandardPreset(layout)
-            }
-            .disabled(
-                model.displays.count != layout.screenCount
-            )
-            .help(
-                model.displays.count == layout.screenCount
-                    ? ""
-                    : "Needs \(layout.screenCount) connected "
-                        + "screen(s)."
-            )
-        }
-        .padding(.vertical, 2)
     }
 
     // MARK: - Saved profiles (#36)
@@ -233,31 +178,5 @@ struct ProfilesSection: View {
                 }
             }
         }
-    }
-}
-
-/// A mini layout diagram (#68 §3.15): one tile per space, each
-/// carrying its mode's glyph (§6.3), so choosing a preset stops
-/// requiring prose alone.
-struct PresetThumbnail: View {
-    let layout: StandardLayout
-
-    var body: some View {
-        HStack(spacing: 3) {
-            ForEach(1...layout.spaceCount, id: \.self) { n in
-                let mode =
-                    layout.spaceModes[SpaceID(n)] ?? .bsp
-                RoundedRectangle(cornerRadius: 3)
-                    .fill(Color.accentColor.opacity(0.12))
-                    .frame(width: 24, height: 18)
-                    .overlay {
-                        Image(systemName: mode.glyph)
-                            .font(.system(size: 9))
-                            .foregroundStyle(.secondary)
-                    }
-                    .help("Space \(n): \(mode.displayName)")
-            }
-        }
-        .padding(.top, 1)
     }
 }
